:hammer: add new shared ui-components project
