+++ package.json
@@ -59,6 +59,7 @@
         "@handsontable/react": "^12.3.3",
         "@ourworldindata/core-table": "workspace:^",
         "@ourworldindata/grapher": "workspace:^",
+        "@ourworldindata/ui-components": "workspace:^",
         "@ourworldindata/utils": "workspace:^",
         "@react-awesome-query-builder/antd": "^6.1.1",
         "@sinclair/typebox": "^0.28.5",
