+++ packages/@ourworldindata/grapher/package.json
@@ -27,6 +27,7 @@
     "@fortawesome/free-solid-svg-icons": "^6.4.2",
     "@fortawesome/react-fontawesome": "^0.2.0",
     "@ourworldindata/core-table": "workspace:^",
+    "@ourworldindata/ui-components": "workspace:^",
     "@ourworldindata/utils": "workspace:^",
     "@tippyjs/react": "^4.2.6",
     "classnames": "^2.3.1",
