+++ packages/@ourworldindata/grapher/src/horizontalColorLegend/HorizontalColorLegends.tsx
@@ -15,8 +15,8 @@ import {
     Color,
     HorizontalAlign,
     VerticalAlign,
-    TextWrap,
 } from "@ourworldindata/utils"
+import { TextWrap } from "@ourworldindata/ui-components"
 import {
     ColorScaleBin,
     NumericBin,
