+++ packages/@ourworldindata/grapher/src/modal/DownloadModal.tsx
@@ -7,9 +7,11 @@ import {
     isEmpty,
     triggerDownloadFromBlob,
     triggerDownloadFromUrl,
+} from "@ourworldindata/utils"
+import {
     MarkdownTextWrap,
     sumTextWrapHeights,
-} from "@ourworldindata/utils"
+} from "@ourworldindata/ui-components"
 import { LoadingIndicator } from "../loadingIndicator/LoadingIndicator"
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome/index.js"
 import { faDownload, faInfoCircle } from "@fortawesome/free-solid-svg-icons"
