+++ packages/@ourworldindata/ui-components/src/MarkdownTextWrap/MarkdownTextWrap.tsx
@@ -1,8 +1,15 @@
 import React, { CSSProperties } from "react"
 import { computed } from "mobx"
 import { EveryMarkdownChildNode, MarkdownRoot, mdParser } from "./parser.js"
-import { excludeUndefined, last, sum, sumBy, imemo, max } from "../Util.js"
-import { Bounds, FontFamily } from "../Bounds.js"
+import {
+    excludeUndefined,
+    last,
+    sum,
+    sumBy,
+    imemo,
+    max,
+} from "@ourworldindata/utils/src/Util.js"
+import { Bounds, FontFamily } from "@ourworldindata/utils/src/Bounds.js"
 import { TextWrap } from "../TextWrap/TextWrap.js"
 
 export interface IRFontParams {
