+++ packages/@ourworldindata/ui-components/src/TextWrap/TextWrap.test.ts
@@ -1,7 +1,7 @@
 #! /usr/bin/env jest
 
 import { TextWrap, shortenForTargetWidth } from "./TextWrap"
-import { Bounds } from "../Bounds.js"
+import { Bounds } from "@ourworldindata/utils/src/Bounds"
 
 const FONT_SIZE = 14
