+++ packages/@ourworldindata/ui-components/tsconfig.json
@@ -1,0 +1,7 @@
+{
+    "extends": "../../../devTools/tsconfigs/tsconfig.base.json",
+    "compilerOptions": {
+        "composite": false,
+        "outDir": "./dist"
+    }
+}
