+++ packages/@ourworldindata/utils/src/index.ts
@@ -596,32 +596,6 @@ export {
 
 export { Tippy, TippyIfInteractive } from "./Tippy.js"
 
-export { TextWrap, shortenForTargetWidth } from "./TextWrap/TextWrap.js"
-
-export {
-    MarkdownTextWrap,
-    sumTextWrapHeights,
-} from "./MarkdownTextWrap/MarkdownTextWrap.js"
-
-export { SimpleMarkdownText } from "./SimpleMarkdownText.js"
-
-export {
-    extractDetailsFromSyntax,
-    mdParser,
-    type MarkdownRoot,
-    type EveryMarkdownChildNode,
-    type EveryMarkdownNode,
-    type EveryMarkdownRootNode,
-} from "./MarkdownTextWrap/parser.js"
-
-export {
-    getLinkType,
-    getUrlTarget,
-    checkIsInternalLink,
-    convertHeadingTextToId,
-    markdownToEnrichedTextBlock,
-} from "./GdocsUtils.js"
-
 export {
     ALL_CHARTS_ID,
     LICENSE_ID,
